installd: support the storage_area directory
We are adding the ability for apps to create "storage areas", which are
transparently encrypted directories protected by screen lock bound keys.
Due to the semantics of filesystem encryption, these directories can't
be located in CE storage and will instead be in a new top-level data
directory, /data/storage_area.  Specifically, an app's storage areas
will be in /data/storage_area/$userId/$packageName.

Therefore, this CL makes the 'installd' changes needed to support this
new app data directory:

- Create and destroy /data/storage_area/$userId/$packageName when an
  app's data directories are created and destroyed.

- Count /data/storage_area/$userId in user disk usage stats, and count
  /data/storage_area/$userId/$packageName in app disk usage stats.

Bug: 280502317
Test: TODO
Ignore-AOSP-First: other changes in the topic need to go to internal
                   master first due to merge conflicts.
Change-Id: I896828644d8807b7df360d592be18de14e0c8dfc

Adding destruction of app-specific storage area when an app is uninstalled

Change-Id: Ic876b1afb3fb41417e846b6328a03223a229bf5f

Fixing typo

Change-Id: I5db1c5568188e1b5c82d3eb04e9b405907ac79c3