libbinder: thread count startThreadPool spawn
When startThreadPool is called, it spawns a thread which is
not counted as part of the lazy kernel-started threads.

This was discovered in fuzzers, and the test is updated.

Bug: 286237215
Fixes: 309885877
Test: binderLibTest
Change-Id: Ib0fa4484576f9d296b8f57f32ae536b17e5c6497
Merged-In: Ib0fa4484576f9d296b8f57f32ae536b17e5c6497
(cherry picked from commit 3e9debc9b6054336d3a54272237aa3726ff53a9d)