binderThroughputTest: compute percentiles exactly
Instead of using buckets to compute the percentiles, store all of the
data and compute percentiles exactly by sorting.

This avoids cases where the computed maximum latency is much much larger
than the average latency. For example, with `-w 64 -i 100000 -p -t`, I
got an average latency of 0.47ms, but the worst latency is 243ms. This
puts everything below 1.8ms in the same bucket.

Bug: 293593894
Test: Manually running the benchmark outputs the correct values
Change-Id: Ib215e4fb503e053294525d944f045555aadae97f
Signed-off-by: Alice Ryhl <aliceryhl@google.com>