binderRpcWireProtocolTest: remove experimental
We are no longer testing that when the wire protocol version
is not experimental, that the experimental wire protocol
version has no changes.

Due to TARGET_RELEASE, we will never mark the 'current' or
'default' version of the wire protocol to be the experimental
version. Instead, the experimental version would stage various
changes to the wire protocol, and we'd update the default
once we want to stabilize that.

Bug: 305786304
Fixes: 309879176
Test: binderRpcWireProtocolTest on -next
Change-Id: Ie4db164bc00ff5980c2c10f0697a05557c752a3a
(cherry picked from commit a5fec0bbd30c82016b06aedddc2fe6fa6ee93588)