Resume in-progress operation after reboot.
Current code does full restorecon if the parent folder needs the restorecon. Unfortunately, the way it's implemented, it sets the folders' attributes first, and there is no way to know if we finished the process.
Selinux library fix is complicated and might break important things in an non-obvious way. Still, we'll look into optimizing the process in a subsequent CLs, and if benchmarks show the benefits.

This marks the folder with 'security.restorecon_in_progress' xattr while we are restoring selinux context.
If for any reason the process is not completed, we'll redo it next time we enter this method - typically after reboot, on user unlock.

Bug: 308541523
Test: presubmit
Change-Id: I221734b0fe0c90e5d39f358efafa69b8cf0430d7