Revert "Remove redundant calls and correct some comments."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f39bab20-6767-4af1-80b4-ceff85e0b42f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I197919d13c8ba708502a338be2e9cced71c00931