Synthesize cancel event to global monitor specify display id and transform
When we failed to find WindowTarget, we will synthesize a cancel event
to globalMonitor, at this time, we should specify the displayId,
because globalMonitor is created according to the different display,
we didn't specify it before, so we add it now.

Meanwhile, when synthesizing cancel event to globalMonitor, we can't
get its transform because globalMonitor doesn't have a windowHandle
associated with it, just like down, move, up events, we should use
the transform of current display.

Bug: 314702448
Test: atest inputflinger_tests

Change-Id: Ic748082383119b465caa41da8301746fefea2be9
Signed-off-by: Linnan Li <lilinnan@xiaomi.corp-partner.google.com>