Report AVC denials to audit
When an SELinux denial occurs within systemmanager, it is logged
directly by servicemanager. For kernel-raised denials, the auditd daemon
consults any bug_map entry to potentially attach a bug to a denial. For
servicemanager, this extra processing is not in place.

With this commit, servicemanager denials are reported to the audit
system. These will be picked up by auditd and processed similarly to
other denials.

Bug: 235163405
Test: As shell, run `service check vold`; observe denials reported by
      servicemanager are correctly routed and processed by auditd.
Change-Id: I93bb123e56667132d2d79bed5e181295f96f11bb