Fix touchpad hover events being suppressed incorrectly
The touchpad event will report a hover event, which is handled by the 
TouchInputMapper, so the tool type is finger, and when the stylus is 
goes down, the touchpad's hover events will all be suppressed, but as
long as it doesn't report an up event, it will be suppressed forever,
even if the stylus is up.
Here we add logic for handling touchpad with action as hover and tool 
type as finger.

Change-Id: Icbaba488f95fff81d9a9c99a44d286966a096e20
Signed-off-by: Linnan Li <lilinnan@xiaomi.corp-partner.google.com>