DO NOT SUBMIT: libbinder: track local refs sent to bps
This is a heuristic test, showing which local objects
are still floating around after a proxy is destroyed.

In many cases, this will show spam/unnecessary data
because the "leaked" refs will be deleted shortly
after. The deletion and promotion check could be
moved to another thread in order to prevent this.

This is also only tested with the following patch
which ensures that BpBinder interface descriptors
will always be available. Without this patch, this
change will likely crash or deadlock:
https://android-review.git.corp.google.com/c/platform/frameworks/native/+/2753215

Bug: N/A
Test: N/A

Change-Id: I6b05eb31cb9e60e7bb0d8a533d1e9985549ad93f