Define a backing type for ADataSpace.
This enum is used in the ABI for the SurfaceControl APIs, but that's
not safe without a defined backing type.

If I've guessed the existing backing type wrong, the ABI checker will
flag this as an ABI break. If the ABI checker is quiet then I've
guessed the inferred backing type correctly.

Bug: None
Test: treehugger
Change-Id: Ifa52f69abacc370e91fd9a358f24a011fdbf0322