Add another pragma to disable another bunch of warnings.
I suspect someone intended there to be an AIOOBE check in here, but I
can't understand the code well enough to be sure of that, let alone well
enough to actually implement it. So just add this warning to the list of
warnings we're already ignoring.

Also update the generator to take into account the stuff that folks have
manually added to the generated files :-(

Test: treehugger
Change-Id: I9f80315612192d65f1e2710a53a9404757ae34bd