Revert "binder RPC tests timeouts to 10min"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a405096-d3a6-48cc-9baa-87d6230c3531). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id8665e68dd4fd518c7700d34419f2cb1946e1909