Revert "binder RPC tests timeouts to 10min"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/71eb7eb0-76c7-41df-a0f5-ed383d066229). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I72e261a6f0d89475b545d3f4296e2c3727d47869