Revert "binder RPC tests timeouts to 10min"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/09c41b7a-1a2f-4f1f-962a-3f45377e2812). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4deab0d062a4f2b47f1496c42a73daefd59dfc0d