Revert "binder RPC tests timeouts to 10min"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/403767ea-5b53-4fad-b1bd-5784ccd11fad). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I781ccb4db5118a138e62cca658c7537a9821b251