Revert "binder RPC tests timeouts to 10min"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/61ad85e4-37b1-4e41-80ba-8b5089a4e026). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id8734d86c906cebe7d611e1522e7be4903b14d0f