Revert "binder RPC tests timeouts to 10min"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/20015fac-fdd6-4582-8407-8d09ca5e45aa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I811302c048cb9c4292185db89a1739ee0eef5b20