Revert "binder RPC tests timeouts to 10min"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1a7206c0-97c1-4d72-945b-5828a3590cd0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0d826608d2b4bbdd954b9bafa9d0ec428665f9e7