Revert "binder RPC tests timeouts to 10min"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/76e532f0-39b4-459d-bd6f-51c0d7bea7e4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I21b091c22ed3aaa29126e62df79859edf5eb5bbe