Revert "binder RPC tests timeouts to 10min"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a6ce9b64-8230-426a-aabb-5cf4e2735e51). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia68e1474cf9aef52119274213effd11949de7c09