Revert "binder RPC tests timeouts to 10min"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8ebdf021-c4ff-40d5-9af0-eaf7288fd8e4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5b3df72cfe692249d69a5f602ec2e6049460aada