Revert "binder RPC tests timeouts to 10min"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/98cfc33a-e472-411c-8f26-113785714d5f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8f7ccc4182236ff7ca05c37e65b4e51ae088e962