Revert "binder RPC tests timeouts to 10min"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0ef4cc30-49c0-4178-b377-9045f2d7112f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie538f8a9c9a28bdf6514166bb30296ed2a4b7774