Revert "binder RPC tests timeouts to 10min"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e53a6e11-8e14-4de8-a1dc-d392d2353a3a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifcd9635b38cc7d36e5aca190cf6b26c21f2f7215