Revert "binder RPC tests timeouts to 10min"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ac26a44-245a-43eb-8b35-e8cea662baa9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6ecd0b8342dced4bd224fcfcb1c80d593a4c16e3