[DON'T BLOCK] Test ownership migration rules
This CL is created as a best effort to migrate test targets
to the new android ownership model. If you find incorrect or unnecessary
attribution in this CL, please create a separate CL to fix that.

For more details please refer to the link below,
<add g3 doc link>

Bug: 304529413
Test: N/A
Change-Id: Ia29683d8fd857ca136d39166346e921473578ca3
Merged-In: Ia29683d8fd857ca136d39166346e921473578ca3