Update tests to treat multithreaded_present as true
These tests were written or updated as if the multithreaded_present
flag is true, but may run on devices where it is not. Explicitly set the
flag to true so the tests always pass.

Conflicts:
- services/surfaceflinger/tests/unittests/SurfaceFlinger_FoldableTest.cpp
The tests which the internal version of
Ica7a921b5bf01b8eadac78900ac69f854f8a02de fixed in
SurfaceFlinger_FoldableTest.cpp do not yet exist in AOSP, so remove
these changes from the commit.

Bug: 323232586
Test: libsurfaceflinger_unittest
Change-Id: Ica7a921b5bf01b8eadac78900ac69f854f8a02de