Revert "Modify dimming threshold from 0.2 to 0.9."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/80ec1c10-fe55-4699-92b9-1e8554ab8481). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibbeea5de665b19e952ee3f681261852b3d6b2e74