Fix Cache being cleared abnormally
When |mMaxTotalSize < newTotalSize| is established, there are the following situations:

    1. |mTotalSize < mMaxTotalSize / 2|, which means |keySize +
    valueSize > mMaxTotalSize / 2|. Since clean() can only clean the
    cache of |mMaxTotalSize / 2| size, it cannot be added after
    cleaning, so isCleanable() Return false, do not trigger clean().

    2. |mTotalSize == mMaxTotalSize / 2|, in this case isCleanable
    should return true and trigger clean().

    3. |mTotalSize > mMaxTotalSize / 2|, this situation does not mean
    |keySize + valueSize <= mMaxTotalSize / 2|, but isCleanable() will
    return true and trigger clean(), and it cannot be added after
    clean().

So the number 3 above should be a bug. Once |(mMaxTotalSize < newTotalSize) && (mTotalSize > mMaxTotalSize / 2) && (keySize + valueSize > mMaxTotalSize / 2)|, clean() will be triggered, but after clean() is executed It cannot be added, and the Cache with the size of |mTotalSize - mMaxTotalSize / 2| will also be cleared, causing the Cache to become smaller.

Therefore, just determine |keySize + valueSize <= mMaxTotalSize / 2| directly in isCleanable().


Bug: 328687163
Change-Id: Ib7127b41ec681a0b2758be5984c8008b7e0cc81b
Signed-off-by: Liangliang Sui <coolsui.coding@gmail.com>