Revert "libbinder: Include stability.cpp in the Trusty NDK library"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6abf94f1-f720-4485-8a10-b80ebfd91d94). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icfb1a1267cffa0bca72cc0a58285a5e7a8c11c83