Avoid sending structs to driver when exts not supported
Libvulkan should only be including the
VkPhysicalDeviceImageCompressionControlFeaturesEXT and
VkPhysicalDeviceImageCompressionControlSwapchainFeaturesEXT structs
in the pNext chain of vkGetPhysicalDeviceFeatures2() during
vkEnumerateDeviceExtensionProperties() handling if the
VK_EXT_image_compression_control and
VK_EXT_image_compression_control_swapchain device extensions
respectively are supported by the underlying driver.

This change updates the loader to query the underlying driver sooner
and then only include the above structs if the extensions are
supported.

Prior to this change, SwiftShader logspam:

  external/swiftshader/src/Vulkan/VkPhysicalDevice.cpp:666 WARNING:
    UNSUPPORTED: curExtension->sType: 1000338000
  external/swiftshader/src/Vulkan/VkPhysicalDevice.cpp:666 WARNING:
    UNSUPPORTED: curExtension->sType: 1000437000
  (repeatedly)

Bug: b/268351352
Test: launch Cuttlefish and inspect logs
Change-Id: I84185ea0cccad650bd9136a61269100e9b0550e5