Revert "Add libdebugstore for Future Use by ART and Frameworks"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e2a4c144-104a-445f-9f33-0a35b38dca72). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I18a1089152890ae9f7a899c8052ef9dcd73d4e9c