Revert "Update for kernel 6.8 headers."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7f5b7017-8f2e-4821-9532-35e7df1cc5e8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If53798a92e90ae7c3510e7ac0858c20b760215ff